添加扫描页面详细广播包数据解析
