+++ app/build.gradle
@@ -5,10 +5,10 @@ android {
     compileSdkVersion 28
     defaultConfig {
         applicationId "com.example.admin.mybledemo"
-        minSdkVersion 18
+        minSdkVersion 19
         targetSdkVersion 28
-        versionCode 11
-        versionName "3.0.1"
+        versionCode 12
+        versionName "3.0.2"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
     }
@@ -55,6 +55,7 @@ dependencies {
 //    implementation 'cn.com.superLei:blelibrary:3.0.1'
     implementation 'com.android.support:appcompat-v7:28.0.0'
 //    implementation 'com.github.skydoves:colorpickerview:1.0.8'
+    implementation 'com.pgyersdk:sdk:3.0.10'
     implementation 'cn.com.superLei:aop-arms:1.0.4'
     implementation 'com.android.support:recyclerview-v7:28.0.0'
     implementation 'com.android.support:design:28.0.0'
