+++ app/release/output.json
@@ -1,1 +1,1 @@
-[{"outputType":{"type":"APK"},"apkData":{"type":"MAIN","splits":[],"versionCode":11,"versionName":"3.0.1","enabled":true,"outputFile":"BLE-v3.0.1.apk","fullName":"release","baseName":"release"},"path":"BLE-v3.0.1.apk","properties":{}}]
+[{"outputType":{"type":"APK"},"apkData":{"type":"MAIN","splits":[],"versionCode":12,"versionName":"3.0.2","enabled":true,"outputFile":"BLE-v3.0.2.apk","fullName":"release","baseName":"release"},"path":"BLE-v3.0.2.apk","properties":{}}]
