+++ app/src/main/java/com/example/admin/mybledemo/AppFileProvider.java
@@ -1,0 +1,6 @@
+package com.example.admin.mybledemo;
+
+import android.support.v4.content.FileProvider;
+
+public class AppFileProvider extends FileProvider {
+}
