+++ app/src/main/java/com/example/admin/mybledemo/FileProvider7.java
@@ -6,6 +6,7 @@ import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.net.Uri;
 import android.os.Build;
+import android.support.v4.content.FileProvider;
 
 import java.io.File;
 import java.util.List;
@@ -28,8 +29,9 @@ public class FileProvider7 {
 
 
     public static Uri getUriForFile24(Context context, File file) {
-        Uri fileUri = android.support.v4.content.FileProvider.getUriForFile(context,
+        Uri fileUri = FileProvider.getUriForFile(context,
                 context.getPackageName() + ".android7.fileprovider",
+//                context.getPackageName() + ".com.pgyer.provider",
                 file);
         return fileUri;
     }
@@ -81,4 +83,5 @@ public class FileProvider7 {
             context.grantUriPermission(packageName, uri, flag);
         }
     }
+
 }
