+++ app/src/main/java/com/example/admin/mybledemo/ui/IntroducedActivity.java
@@ -1,0 +1,19 @@
+package com.example.admin.mybledemo.ui;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.example.admin.mybledemo.R;
+
+public class IntroducedActivity extends AppCompatActivity {
+
+    private static final String TAG = "IntroducedActivity";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_introduce);
+
+    }
+
+}
